added primitive Item model. See comments.
In ReservesDirect, an item representation is spread across several
tables. I propose that we try to model an item as a row in a single
table, and include columns for the item metadata as well as its
physical status (if any) and content (if any).

Perhaps it will become clear why RD uses so many tables, and I'll see
the error of my ways. :-) But it seems wise to start with the simplest
data model that might work, and extend from there.


git-svn-id: svn://svn.open-ils.org/ILS-Contrib/servres/trunk@47 6d9bc8c9-1ec2-4278-b937-99fde70a366f