Admin options for editing terms, deparments: generic indexes and forms.
The big change in this rev is introduction of the Django 'newforms'
system. Since we're using Genshi (not Django templating), newforms
seemed a bad fit. But I think I've got a fairly nice compromise
between Django and Genshi going on in my generics.py.

See the ModelForm subclasses in views.py: the Index attribute is
novel, and specific to the generics.py system.

There is still more to do here, notably figuring out how to handle
permissions and i18n.

git-svn-id: svn://svn.open-ils.org/ILS-Contrib/servres/trunk@133 6d9bc8c9-1ec2-4278-b937-99fde70a366f