i18n review -- tagged all the English literals, rebuilt POT
There are a couple oddballs that I'm not sure how to handle. For
exmaple, how does one internationalize an HTML literal like "This is a
subproject of <a href="...">Project Conifer</a>."? I don't see
anything in Babel that has the right kind of smarts for this. Must ask
an i18n guru.


git-svn-id: svn://svn.open-ils.org/ILS-Contrib/servres/trunk@135 6d9bc8c9-1ec2-4278-b937-99fde70a366f