style and heading-level changes. Add-new-course now requires an extra perm.
I'm not sure how in practice we will give add-course permissions to
instructors. How do we know which ones are the instructors?

I think our backend for user-information is going to have to grow an
(optional) is_user_an_instructor() test. We can also add an
Instructors group in Django which will give the required permission.


git-svn-id: svn://svn.open-ils.org/ILS-Contrib/servres/trunk@154 6d9bc8c9-1ec2-4278-b937-99fde70a366f