working on Mark Physical Items As Arrived: fuzzy match is working.
The fuzzy-match is in place; not tuned yet, but it's there. It's a
relevance-engine based on Levenshtein-distance comparison of the
title, author (and to a lesser degree the publisher and
pubdate). Ideas for improvements are most welcome.

Note that this version of the code takes a full snapshot of the MARC
record when a Physical Item is requested from the catalogue. So there
are more opportunities for item comparison.

The code is a horrible mess. Much cleanup to do.

git-svn-id: svn://svn.open-ils.org/ILS-Contrib/servres/trunk@258 6d9bc8c9-1ec2-4278-b937-99fde70a366f