Enabled ANON courses. Search and browse results are now limited by user-permissions level.
ANON courses are courses which are open to non-authenticated
users. I'm not sure that this is a good idea yet, but it is useful for
testing.

Browsing will no longer show, for example, courses which the current
user cannot enter. Search will not return items that the current user
cannot view. This is extended to browse-instructors: if the instructor
has no accessible courses, don't bother showing the course.


git-svn-id: svn://svn.open-ils.org/ILS-Contrib/servres/trunk@396 6d9bc8c9-1ec2-4278-b937-99fde70a366f