Changed "Add a new course site" to "Create a new course site".
Betty Ing suggested this, for consistency. I'm not sure that all the
other "Add a new" links should be changed to "Create a new"s,
though. Creating a course site makes sense to me; but you don't create
a term (though you might add a representation of one to the system).
Opinions welcome.


git-svn-id: svn://svn.open-ils.org/ILS-Contrib/servres/trunk@399 6d9bc8c9-1ec2-4278-b937-99fde70a366f