Fix edit-site bug; better save-confirm messages on Permissions; delete groups.
When editing a site (via "Setup"), the owner name was not being set properly,
and would cause the edit to fail if the owner was not changed.

Save-confirmation messages now have a "continue" link on the Permissions page,
making it a bit easier to proceed after making changes.

External groups are now deletable.

git-svn-id: svn://svn.open-ils.org/ILS-Contrib/servres/trunk@1127 6d9bc8c9-1ec2-4278-b937-99fde70a366f