Simplify the Sakai integration somewhat: no 'create/revive site' options.
Instructors can choose to link to an existing reserves list with which
they are associated. If they do so, instructors and students visiting
from Sakai will be taken directly to the linked site. If the
instructor chooses not to link a reserves list, then visitors from
Sakai will be taken to the generic "Browse" page.

git-svn-id: svn://svn.open-ils.org/ILS-Contrib/servres/trunk@1153 6d9bc8c9-1ec2-4278-b937-99fde70a366f