Now using classes, not modules, to implement local integrations.
This is a rather big change, structurally, that lets us define integrations
naturally in terms of other integrations. For example, UWindsor is now a
subclass of EvergreenSite; the code which is Windsor-specific isn't mingled
with the general Evergreen stuff.

There are still many Leddy/Windsor-isms (and Evergreen-isms) throughout the
codebase that need to be sorted out into new integration functions.

git-svn-id: svn://svn.open-ils.org/ILS-Contrib/servres/trunk@1303 6d9bc8c9-1ec2-4278-b937-99fde70a366f