Fix e-resources bug and add note about academic terms to README
Thanks to shadowspar for fixing bug that adds electronic resources from the catalog
as physical items (PHYS) instead of URLs and cleaning up URLs that have additional
paramters after the bib id. Academic terms are hardcoded in the integration class
and a long overdue note has been added to README. The next step is to add an actual
admin interface for this.

Signed-off-by: Art Rhyno <art632000@yahoo.ca>