Add Parts Support
Syrup was developed before the parts mechanism in evergreen was
available. I have left the munging option in place which tries to
deal with call numbers that have parts information embedded in the
call number string. Hopefully, such practices will not be common
anymore since parts is wayyyy more elegant.

Signed-off-by: Art Rhyno <art632000@yahoo.ca>