Group parts and copies but allow editors to have item access
It has been necessary to suppress items so that copies of parts
don't appear that are not on reserve. These changes only display
status information for items identified in syrup, grouping them
for anonymous and member access, but allowing individual editing
for anyone with permissions to change the course site.

Also, one teeny-tiny change to remove the word "Faculty" from the
"How to place materials on reserve" link,

Signed-off-by: Art Rhyno <art632000@yahoo.ca>