Link to earliest available copy
Making syrup more multipart aware opened the door to
linking from the summary page to the full display
of a random copy. The summary page is set to always
show when the next available copy is due and it can
be jarring to see a different due time on the full
display. This should hopefully take care of that
scenario.

Signed-off-by: Art Rhyno <art632000@yahoo.ca>