WIP: naive fix for superpage error in cat_search
This change was successful in working around a change in the
expected object structure returned when trying to create a list of
IDs from an Evergreen catalog search via
open-ils.search.biblio.multiclass.query

Tested on an Evergreen 2.11 system running OpenSRF 2.4. Even if this
is a reasonable approach to the problem, it likely needs more in the
way of conditionals, etc.

Signed-off-by: Jeff Godin <jgodin@tadl.org>